S15.b: decompose cli/src/commands/mesh.ts (1583 → 667 LOC)
Split mesh.ts along natural seams into a fresh cli/src/commands/mesh/
directory, bringing it under §15 800-LOC cap:

- mesh/identity.ts (137 LOC) — MeshIdentity + AES-256-GCM at-rest
  encryption + Ed25519 keypair + AMID derivation + base58 + load/save
- mesh/oauth.ts (94 LOC) — OAuth callback HTTP server + escapeHtml
  (CWE-79) + sanitizeForLog (CWE-117)
- mesh/health.ts (127 LOC) — port + WS health helpers
- mesh/auth.ts (221 LOC) — mesh auth subcommand body
- mesh/promote.ts (409 LOC) — mesh promote subcommand body

Public re-export surface preserved (generateKeypair, base58Encode,
encryptPrivateKey, decryptPrivateKey, checkRegistryHealth,
checkRelayHealth, killProcessesOnPorts, killStaleListeners, type
MeshIdentity); mesh.test.ts (28 tests) passes unchanged. All 454 CLI
tests pass; tsc/lint/build clean.

No behavioral change: every helper and subcommand action body moved
verbatim. File mode bits, OAuth bind address, machine-bound key
derivation, and registry/relay health logic all preserved.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-mesh-cli.md
